% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/railtrails.R
\docType{data}
\name{railtrails}
\alias{railtrails}
\title{Rail-trail trail data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11420 rows and 8 columns.}
\source{
https://www.traillink.com/
}
\usage{
railtrails
}
\description{
Information and reviews for rail-trails in the United States
}
\examples{
railtrails

# to expand vector of review ratings in raw_reviews column:
library(tidyr)
railtrails <- railtrails \%>\% unnest(raw_reviews)
}
\keyword{datasets}
